First Draft
This is my first draft of the hangman game, I am going to work on now creating a game over (when you guess more than 6 bad guesses) and a game win (when you guess the phrase or word correctly without exceeding 6 bad guesses). I am also going to try and have a random word generator to make the game a better representation of hangman.